Add a request verification service to verify incoming requests
The SignedRequest value-object can be constructed from PHP globals or
by passing in the desired data by hand.

Construct the RequestValidator service with your account's signing key
and pass the SignedRequest object to test freshness of the request
and whether the data in the request matches the signature.

Because we use the PHP hashing functions, I've included a polyfill for the
functions that are not available in older PHP versions.